tests/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
atomicwrites==1.4.0
    # via pytest
attrs==21.2.0
    # via pytest
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
colorama==0.4.4
    # via pytest
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
psycopg2-binary==2.9.2
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via -r requirements.in
requests==2.26.0
    # via -r requirements.in
toml==0.10.2
    # via pytest
urllib3==1.26.7
    # via requests
